Relax encoding regex (#7)
* Relax encoding regex

I had an epub which was using additional attributes in the `xml` tag (in my case it was `<?xml version="1.0" encoding="UTF-8" standalone="no"?>`). The regex wasn’t matching it and therefore added an additional `xml` tag which made the file invalid. 

Since we don’t care if there are any additional attributes, this change relaxes the regex so it matches tags like the above. 

Thanks for the super-handy tool! I use it all the time.

* Use non-greedy matching for xml tag regex